fix(llm-catalog,cli): fix publish + Docker build breaks surfaced by CI
- packages/llm-catalog: inline auth-requirements.ts's content back into
  index.ts. The package publishes dist/ via tsc with
  moduleResolution:"Bundler" (tsconfig.build.json), which does not emit
  the explicit .js extensions plain Node ESM needs on relative imports —
  a separate ./auth-requirements module broke the SDK's install-smoke test
  (packages/sdk/scripts/smoke-install.mjs: Cannot find module
  '.../dist/auth-requirements' importing the actually-published tarball).
  Every other export in this package has always lived in this one file for
  exactly this reason; verified locally via
  `pnpm --filter @kortix/sdk run smoke:install` against the real packed
  tarball (previously reproduced the failure, now passes).
- apps/cli/src/index.ts: wire `providers` into the command dispatcher —
  apps/cli/src/commands/providers.ts existed but was never imported/
  registered, so `kortix providers ...` was unreachable before this change
  too (pre-existing gap, found while validating the CLI parity fix end to
  end). Added to KNOWN_COMMANDS and the top-level help listing.
- apps/api/Dockerfile, apps/sandbox/Dockerfile: both assemble a minimal
  throwaway workspace to compile the CLI binary and only COPY a curated
  subset of packages/*. @kortix/llm-catalog is now a real dependency of
  apps/cli (needed for the auth-requirements contract) — added
  `COPY packages/llm-catalog ./packages/llm-catalog` to both. Verified by
  building the `sandbox-cli` stage locally with Docker (previously failed
  with "Workspace dependency @kortix/llm-catalog not found", now succeeds).
